Added a test for MasterInfo.
We want to test that when the master gets initialized, the
MasterInfo that it creates (and gets ultimately serialized to
ZooKeeper) has the correct information in its Address field.

Currently, the test hangs when run in after another
MasterZooKeeperTest, due to some issues around the recovery.

Jira: MESOS-2736

Review: https://reviews.apache.org/r/36807