Published MasterInfo to ZK in JSON format.
As part of the support for non-libmesos linked client to
interact with the HTTP API, we also provide the Master information
stored in ZooKeeper in a way that is deserializable also for
clients that do not necessarily know how to decode a MasterInfo
protocol buffer.

This patch publishes the data in JSON format and has been tested,
summarily, with 2 masters running this patch alongside a 0.23 Master and
one each of 0.24 and 0.23 Agent nodes.

Please note this patch does not require specific unit testing, as the functionality
is already widely tested in existing `ZooKeeper*` tests, as well as `MasterDetector*` ones.

Review: https://reviews.apache.org/r/36900