Add explicit compiler check for CMake Windows builds.
MesosConfigure.cmake will currently allow users to try to compile
Mesos with any compiler and any version on the Windows platform. But,
the for now we will only support MSVC 1900 or later.

Our solution is to remove the possibility of compiling on anything
other than MSVC 1900. This commit will introduce explicit warnings
when users attempt to compile with anything else.

Review: https://reviews.apache.org/r/36964