Disable CMake support for building process lib and tests on Windows.
We're in the process of adding Windows support for all these
libraries.

This commit will temporarily disable CMake from attempting to compile
these projects on Windows. Once we add support in the code, we will
enable them in the CMake build system as well. This is part of our
transition towards a CMake that incrementally builds up Windows
support.

Review: https://reviews.apache.org/r/36966