Add path configuration for pulling down curl dependency in Windows.
Windows does not have a good default package manager, and so the
overhead for installing dependencies that CMake historically assumes
are simply present (such as curl) is a _huge_ chore.

A simple fix is to simply use CMake to fetch these dependencies. This
commit will introduce the path configuration logic that will make this
possible for the curl dependency. In subsequent commits we will add
the CMake code download the curl dependency in the Stout project.

Review: https://reviews.apache.org/r/36968