Implemented a LRU entry selection criteria for cache eviction.
A linked list is used to keep cache entries in LRU-to-MRU order. Each time an existing cache entry is requested, it is moved to the back of the list. During cache eviction entries are removed from the front of the list until enough cache space can be freed.

Review: https://reviews.apache.org/r/36773