Updated submitting-a-patch.md with bootstrap command.
When developer is using git to clone the mesos source code and
want to build & submit a patch, s/he needs to run ./bootstrap
first to generate some files such as configure, .gitignore etc.
But the document do not include such instructions.

This patch is adding bootstrap to the guidelines.

Review: https://reviews.apache.org/r/36472