Updated slave task label decorator hook to pass in ExecutorInfo.
If the task being launched has a command executor, there is no way for
the hook to determine the executor-id for that task. This update calls
the hook _after_ the ExecutorInfo has been created and thus is able to
pass in ExecutorInfo to the label decorator.

Review: https://reviews.apache.org/r/37228