Updated authorizer to allow for modularized implementations.
Splits and updates the original declaration of the `Authorizer` into
its interface and a default implementation, the `LocalAuthorizer`.

Following the pattern of the modularized `Authenticator`, it generates
a default constructor which is required when writing a `TYPED_TEST` in
a follow up patch. Additionally, an initialize method has been added,
needed for passing in the current ACL definitions as provided via
flags.

Other changes are just updates to allow for compilation.

Review: https://reviews.apache.org/r/36048