Added "labels" to FrameworkInfo.
This is intended to support frameworks that want to offer capabilities to the
rest of the cluster (e.g., storage or some arbitrary third-party service). We
want processes running in the cluster to be able to discover these capabilities;
however, we don't want to commit to a fixed set of capabilities or how those
capabilities should be represented. Hence, this commit represents this
information using freeform key-value pairs, similar to the labels mechanism
already in use elsewhere.

Jira: MESOS-2841

Review: https://reviews.apache.org/r/37443