Add --ip_discovery_command to Agent.
Similarly to the Master, we want to enable the Agent too to be
initialized with an IP address that is "discovered" by an arbitrary
script.

This is particularly useful for Mesos deployments on public (or
private) clouds where DNS resolution is either limited or not
feasible.

Jira: MESOS-3154

Review: https://reviews.apache.org/r/37457