Fixed bug in fetcher.cpp when HDFS client is not available.
When trying to download an artifact with the Hadoop client,
and the client is not `available()` we correctly return a
false value, but then in the error message, we tried to
make a call to `Try::error` which failed and crashed Master.

This fixes it.

Review: https://reviews.apache.org/r/37584