Update README.md
Hi,

I added section "Books" and a link to the Mesos in Action written by Roger Ignazio for Manning Publications. We believe this book is a great resource and all members of community will benefit from this information. If you find this inappropriate, please let me know where I can publish it.