Added containerId to ResourceUsage to enable the QoS controller to target a container.
We should ensure that we are addressing the container which the QoS
controller intended to kill. Without this check, we may run into a
scenario where the executor has terminated and one with the same id has
started in the interim i.e. running in a different container than the
one the QoS controller targeted.

This most likely requires us to add containerId to the ResourceUsage
message and encode the containerID in the QoS Correction message.

Review: https://reviews.apache.org/r/38253