Fixed race in hook self-message loop and reenabled VerifySlaveLaunchExecutorHook test
Coordinating events across the library border is hard as we want to
avoid exporting additional symbols between the test and the module code.
To migitate this, the VerifySlaveLaunchExecutorHook used a technique
where it creates a libprocess actors in-place and sends a message to
itself. This can be caught by a message filter in the shared libprocess
instance and the test code can synchronize over this, to make sure
certain module code was executed.

However, the in-place actor could (potentially) shutdown before the
message was received (and thus, didn't execute the filter).

This patch installs a message handler in the in-place actor and only
shuts down the actors when the message has been received.

Review: https://reviews.apache.org/r/38574