Pass LIBPROCESS_IP even when executor environment is specified.
If DNS is not available on the agent node and a task is launched which
explicitly specifies the executor's environment, LIBPROCESS_IP will not
be passed through and the default hostname lookup after spawning the
executor process will throw an error. This patch alters the agent to
always pass LIBPROCESS_IP, even when the executor environment is
specified.

Review: https://reviews.apache.org/r/39152