Added code that appends the fetcher log to the agent log upon fetcher failure.
Adds an onFailed() clause to the inspection of the fetcher subprocess run. This clause copies the fetcher log from <task sandbox>/stderr and appends it to the agent log.

This is to facilitate debugging spurious fetch failures in production or CI.

Similar, but not the same: https://reviews.apache.org/r/37813/ (see MESOS-3743 for an explanation).

Review: https://reviews.apache.org/r/39338