Documented messages in messages.proto.
A brief summary of each message was added.

For messages with an associated Event/Call API object, a reference to
the object was added.
Additionally, there is a great deal of documentation overlap between
these messages and the comments in mesos/scheduler.hpp and
v1/scheduler/scheduler.proto.  Where necessary, some notes were added
about the backwards compatibility of messages which are not instantiated
in the code base.

Review: https://reviews.apache.org/r/39457