Moved the framework/executor lookup in the Agent's executor endpoint.
Refactored the `api/v1/executor` endpoint to look up the framework and
executor before allowing subscription. This is expected since the slave
stores the framework and executor during the launch process. Updated the
tests to first spawn a dummy `cat` process and then make a `Subscribe`
call to ensure that the Agent knows about the framework / executor and
not reject it with a `BadRequest`.

Review: https://reviews.apache.org/r/38875