Added function that verifies prerequisites for using Linux launcher.
We recently switched to using LinuxLauncher by default. This causes
problems with running the slave in a Docker container.

This patch introduces a function for checking whether the machine is
suited for running Linux launcher. For the time being it checks whether
the freezer cgroup subsystem enabled, as the Linux launcher evolves we
might want to add more checks.

Review: https://reviews.apache.org/r/39604