Fix failed test LimitedCpuIsolatorTest.ROOT_CGROUPS_Pids_and_Tids
This test is to isolate one single process with one thread in cgroup, and checks
the number of processes and threads in the cgroup before/after isolation. The
previous test case uses "sh -c "while true; do sleep 1; done;"" as the process,
however, it periodically forks a child process "sleep 1" in cgroup causing the
test failure (expected number of process/thread should be 1, but here is 2).
This patch is to issue a "cat" command, so there will be only one process in
cgroup.

Review: https://reviews.apache.org/r/38454