Mesos: Corrected style of Makefiles.
This is not a full standartization of Makefile format, however, it is
a step towards consistency. This patch focuses mainly on the order of
files and corrects obvious issues. These are guidelines behind the
change:

- Files in the same folder should appear in lexicographical order.
- Subfolders should appear in lexicographical order either after files
or mixed with them (depending on the current consensus in the file).
- Though tabs are preferred, spaces are not replaced if they are
current consensus in the file or current section.

Review: https://reviews.apache.org/r/40011