Added a paths:: helper for executor HTTP marker files.
This change adds the relevant functionality to the paths utilities
to store a marker file to denote HTTP based executors. We will
create the file when `checkpointing` is enabled as part of handling
the `Subscribe` request. This will then used by the agent when
recovering to ascertain if the executor is HTTP-based.

Review: https://reviews.apache.org/r/38876