Fixed a bug with HTTP pipelining in libprocess.
Response futures were being added to the pipeline not when the
requests arrived but when the event was visited. This can
result in out of order responses.

This patch fixes that problem enforcing that every response is
sent in the order the corresponding request arrived.

Review: https://reviews.apache.org/r/39276