Introduced an Authenticator interface and an AuthenticationRouter in libprocess.
The Authenticator interface allows us to implement different
authenticators based on the scheme (e.g. Basic, Digest, SPNEGO).
The AuthenticationRouter manages the authentication realms and
the mapping from endpoints to realms. It is then used by the
ProcessManager to route requests to the authenticator for the
realm, if applicable.

Review: https://reviews.apache.org/r/37999