Always create non-IP egress filters in port mapping isolator.
When rolling out the new flag --egress_unique_flow_per_container, I
noticed, on some slaves, only IP egress filters were created as
expected, the reset were not. Looking at the code, it looks like we
skipped the creation if this is not the first container we create, this
is wrong for this case, because egress filters were not created for
previous containers yet. We should always create them and ignore error
if they exist.

Review: https://reviews.apache.org/r/39490