Added functionality for Subscribe/Subscribed workflow for HTTP executors.
This change adds the functionality for executors to `Subscribe` via the
`api/v1/executor` endpoint. It also stores a marker file as part of the
`Subscribe` call if framework `checkpointing` is enabled. This can then
be used by the agent when recovering to wait for reconnecting back with
the executor.

Review: https://reviews.apache.org/r/38877