Aggregated as member Http in Master.
Before this change the Http object was copied in to each HTTP routing
lambda. If state were to be introduced, this could create hard to
debug issues. Aggregating a single instance as a member of Master
prevents these issues and simplifies access to communal code paths
such as `authenticate`.

Review: https://reviews.apache.org/r/41211