Added a benchmark to simulate frameworks declining offers.
This benchmark starts a number of slaves and frameworks, then cycles
the allocator. On each allocation pass, the frameworks decline all
the offers. This leads to increasing numbers of refusal filters in
the allocator, and the allocation slows down. After around 200
cycles the slowdown increases out of proportion.

Review: https://reviews.apache.org/r/38649