Reworded NOTE in Examples for Getting Started doc.
The previous wording was a bit confusing. The new text makes it clear
that the test-framework binaries have been generated by a call to
`make check` as described in the 'Building Mesos' section.

Also, I changed the text from being a footnote to being actual text
inline with the rest of the text in this section. The note is relevant
enough to the text above it to warrant such a change.

Review: https://reviews.apache.org/r/41363/