Jenkins: Updated docker build script for centos 7.
Jenkins builds are consistently failing for centos 7, with the
failure:

checking value of Java system property 'java.home'...
/usr/lib/jvm/java-1.8.0-openjdk-1.8.0.65-3.b17.el7.x86_64/jre
configure: error: could not guess JAVA_HOME

They also fail early on during 'bootstrap' with a missing 'which'
command.

The solution is to update support/docker_build.sh to install 'which' as
well as make sure the proper versions of java are installed during the
installation process.

The problem with java is that we install maven BEFORE installing
java-1.7.0-openjdk-devel, causing maven to pull in a dependency on
java-1.8.0-openjdk. This causes problems with finding the proper
java.home in our mesos/configure script because of the mismatch between
the most up to date jre (1.8.0) and the most up to date development
tools (1.7.0). We can either update the script to pull in the 1.8 devel
tools or move our dependence on maven until AFTER our installation of
java-1.7.0-openjdk-devel.

This commit takes the approach of updating to explicitly install the
java-1.8.0 tools.

Review: https://reviews.apache.org/r/41475/