Added support for implicit roles.
Changed the behavior of the master when the `--roles` flag is NOT
specified. Previously, this would allow only the `*` role to be used.
Now, omitting `--roles` means that any role can be used. This is called
"implicit roles". Configuring which principals can perform operations as
which roles should be done using ACLs in the authorization system.

Note that this changes the behavior of the system when `--roles` is not
specified. This is likely acceptable: if the operator didn't specify
`--roles` in prior versions of Mesos, they were likely not using roles
or authorization at that time.

Another minor behavioral change is that the "/roles" endpoint will now
only return results for currently "active" roles (those with one or more
registered frameworks).

The `--roles` flag is now considered deprecated and will be removed in a
future version of Mesos.

Review: https://reviews.apache.org/r/41075/