Fixed a documentation bug in containerizer internals doc.
The current document delves upon the behavior of using the default
executor i.e. `mesos-docker-executor` on whether the task has `TaskInfo`
set or not. This is slightly confusing as any task in Mesos is specified
via the `TaskInfo` protobuf message.

This change corrects this and outlines the difference in behavior on
whether the `Task` specifies an executor (custom executor) or specifies
a command (command executor).

Review: https://reviews.apache.org/r/41697/