Cleaned up configuration.md.
Edits configuration.md for style and consistency. Makes consistent use
of <code></code> for commands, flags, and flag values.

Edits the libprocess flags for consistency and proper use of `=DIR` or
`[=DIR]` depending on whether or not the directory is an optional
parameter.

Note that the libprocess `--with-*` flags have two different
descriptions depending upon whether or not the library is bundled with
Mesos.

Review: https://reviews.apache.org/r/39923/