Used only exactly representable floats in test calculation.
In the IEEE floating point standard representation usually only values
where the fractional part of the normalized form is a multiple of 2^n
can be represented exactly; a value of 0.33 cannot be represented
exactly.  This can lead to an incorrectly failing tests where the
calculated value does not match the expectation.

Instead switch to a value of 0.25 which can be represented exactly.

Review: https://reviews.apache.org/r/41872/