Using 'git rev-parse --git-common-dir' in post-reviews.py.
The post-reviews.py script has similar problems to the bootstrap script
in assuming that .git is a folder in the top-level directory of the git
repo.  This may not be true when mesos is included as a submodule in
another project. Use 'git rev-parse --git-common-dir' instead.

Review: https://reviews.apache.org/r/41244/