Added flags to update summary / desc for reviews.
Previously if you amended a commit message and posted it to reviewboard
via the support/post_reviews.py script, then the summary / description
on reviewboard would not be updated. The only way to update these was
to modify them directly on the webpage. However, with a simple
directive in the .reviewboardrc file we can force the review's summary
/ description to be updated to the text in the commit message.
This commit adds these directives to the default .reviewboardrc.

Review: https://reviews.apache.org/r/41411/