Updated post-reviews.py to strip review URL in RR description.
Previously, when setting GUESS_FIELDS in .reviewboardrc to automatically
update the description message when posting a review, the description
would include the review URL in the description message. This was due
to the fact that GUESS_FILEDS simply copies the entirety of the commit
message into the RR description field on reviewboard.
Including this line is both redundant, and, in some cases, has been
found to appear (at least) twice in the description message instead of
just once.

This commit ensures that the commit message posted when GUESS_FIELDS is
set will have the review URL stripped before posting.

Review: https://reviews.apache.org/r/42266/