Changed 'ReservationInfo.principal' from required to optional.
In order to allow dynamic reservation without a principal, this field is
being changed to optional. However, the current patch alters the master
to invalidate any reserve operations that do not set this field. After a
deprecation cycle, the master will allow the field to be unset.

Review: https://reviews.apache.org/r/42334/