Added dynamic reservation test with no authentication.
Currently, we do not allow dynamic reservations without a principal.
The `ReservationInfo.principal` field is currently being migrated to
`optional`, and in 0.28.0 we plan to begin allowing dynamic reservations
without a principal. This test was added to verify that currently,
framework reserve operations without a principal will fail as expected.

Review: https://reviews.apache.org/r/42361/