Fixed whitespace errors in help strings for mesos/src.
Previously some of the help strings included a blank space at the end of
some of their lines of text. This was causing problems when using these
strings to generate .md files and then attempting to commit them back to
the repo.

This commit removes this unnecessary whitespace.

Review: https://reviews.apache.org/r/42787/