Added variadic template for process::await.
Previously, templates only existed to allow await to take either a
std::list of Futures, *exactly* 2 Futures as arguments, or *exactly* 3
Futures as arguments.

This commit removes the 2 and 3 argument templates and replaces them
with a variadic template that accepts an arbitrary number of Futures as
arguments.

Review: https://reviews.apache.org/r/42956/