Changed ZooKeeper reconnection logic to retry more aggressively.
The previous implementation of `GroupProcess` tried to establish a
single ZooKeeper connection on startup, but didn't attempt to retry.
ZooKeeper will retry internally, but it only retries by attempting to
reconnect to a list of previously resolved IPs; it doesn't attempt to
re-resolve those IPs to pickup updates to DNS configuration. Because
DNS configuration can be quite dynamic, we now close the current Zk
handle and open a new one if we've seen a successful `zookeeper_init`
but haven't been connected within the ZooKeeper session timeout.

Review: https://reviews.apache.org/r/42988/