Made sure the container launcher terminated before we leave the test.
Waiting for the launch result is not enough as it is unrelated to
whether the launched container has terminated. If we do not wait for
that the global teardown might find it still running and fail the tests
(this can happen if we e.g., execute this test in isolation).

Also aline the name of the launch result to follow the canonical
scheme.

Review: https://reviews.apache.org/r/42247/