Added a process ID argument to Slave constructor.
This change modifies the existing `Slave` constructor to take in the process ID
string as an argument. This is necessary for testing recovery of HTTP based
executors. (Previously, every invocation of `StartSlave` used to generate a new
ID via `ID::generate`). There was no process delegate set via
`process::initialize` that led to problems for the existing HTTP based executor
to connect back to the slave.

Also, modified the tests to introduce a new `StartSlave` function that takes
this process ID as argument.

Review: https://reviews.apache.org/r/42181/