Enable mesos tests installation.
This patch enables the installation mesos-tests and its dependencies
and helper tool. The goal is to allow operators to build a separate
test package that can be run at deployment time to verify that Mesos
works in the deployment environment.

Since the build directory is searched first, to run it on a host
that has a build tree, you need to specify a non-existent tree:

~ $ $PREFIX/libexec/mesos/tests/mesos-tests --build_dir=/none

\- Add --enable-tests-install
\- Fix mesos-tests gmock dependencies
\- Optionally install tests, helpers and test modules
\- Add utility helpers to find various test resources

Review: https://reviews.apache.org/r/40553/