Fixed flakiness in SlaveRecoveryTest/0.CleanupHTTPExecutor.
This change fixes the flakiness in this test. The issue was a race between the
`connected` callback being called before we did `process::spawn` to start the
process.

The details of the race that lead to the failure are as follows:
- We started the executor library inside the constructor of `TestExecutor`. The
  callback function did `process::defer(self(), &Self::connected)`
- The `connected` callback can be invoked by the Executor library before we got
  a chance to actually invoke `process::spawn` on the `TestExecutor` process
itself. This in can turn lead to the `dispatch` being silently dropped.

This change now starts the library inside the `initialize` function that is
gurranteed to be called after `process::spawn` is invoked.

Review: https://reviews.apache.org/r/43285/