Added examples for documentation on master/slave state.json.
We are working on a more standardized way of representing (and thus
more easily documenting) our internal JSON objects. Until we have
this, there is no point in trying to exhaustively document the
state.json (or any other JSON) endpoints. Including an example and
noting that it is not exhaustive is better than it was before though.

Review: https://reviews.apache.org/r/43618/