Documented that labels should not contain duplicate key-value pairs.
The implementation of the equality operator for `Labels` is buggy for
labels that contain duplicates. We might want to revisit fixing the
implementation of that operator (which might be expensive; MESOS-4445),
but in the short-term we should document that duplicates should not be
specified.

Review: https://reviews.apache.org/r/43616/