Introduced a TASK_KILLING state.
TASK_KILLING can be used to signify that the kill request has been
received by the executor, but the task is not yet killed. This is
similar to how TASK_STARTING indicates the launch request has been
received by the executor, but the task is not yet launched.

This new state will be guarded by a framework capability in order
to ensure that we do not break older frameworks.

Review: https://reviews.apache.org/r/43487/