Allowed dynamic reservation without a principal.
The `ReservationInfo.principal` field has been migrated to `optional`,
which means we can now allow dynamic reservation and unreservation
without a principal. This allows the use of the `/reserve` and
`/unreserve` HTTP endpoints when HTTP authentication is disabled.

Note that we still require that frameworks/operators set the
`ReservationInfo.principal` field to match their own principal,
if present. It may be desirable to remove this requirement;
this improvement is tracked in MESOS-4696.

Review: https://reviews.apache.org/r/43639/