Cleaned up MesosSchedulerDriver shutdown in unit tests.
For consistency, we should do `driver.stop(); driver.join();` if a
`MesosSchedulerDriver` has been started by the unit test. The destructor
for `MesosSchedulerDriver` will do something _similar_, so the
consequence of omitting these calls is not dire, but it seems safer to
be consistent between all the test cases.

Review: https://reviews.apache.org/r/42877/