Updated docs for reservation, volumes, and authZ.
This updates the authorization documentation to include the new `roles`
object for the `CreateVolume` and `ReserveResources` ACLs. The docs for
persistent volumes and dynamic reservations are also updated to reflect
the new authorization behavior. A note has been added to `upgrades.md`
detailing the impact of these changes on upgrades.

Review: https://reviews.apache.org/r/43800/