Added 'Updating the wiki' step to the release guide.
In the recent Mesos Community Sync, it was reminded that we should be keeping
the [Mesos Release
Planning](https://cwiki.apache.org/confluence/display/MESOS/Mesos+Release+Planning)
wiki page up to date.  This patch adds this step to the release guide so that we
don't forget.

Review: https://reviews.apache.org/r/44108/