Added support for pipelining calls to the scheduler library.
Previously, the scheduler library used to chain calls on previous call
responses. This was inherently slow. This change adds support for pipelining all
calls to the master on a single connection via the `http::Connection`
abstraction in libprocess.

This change also adds support for handling various error scenarios when we
notice a disconnection instead of just relying on the master detector for
invoking the `disconnected` callback.

Review: https://reviews.apache.org/r/43662/