Windows: Launch: Removed `rootfs` launcher flag, preventing `chroot`.
`chroot` does not exist on Windows. Unfortunately, the launcher also
depends on it. In this commit, we removed Windows support for the
launcher flag `rootfs`, which controls whether we use `chroot` in the
launcher. This allows us to divest ourselves of `chroot` altogether on
Windows.

Review: https://reviews.apache.org/r/43904/