Windows: Removed `user` launcher flag, preventing `su`.
`su` does not exist on Windows. Unfortunately, the launcher also depends
on it. In this commit, we remove Windows support for the launcher flag
`user`, which controls whether we use `su` in the launcher. This
allows us to divest ourselves of `su` altogether on Windows.

Review: https://reviews.apache.org/r/43905/