Removed internal usage of deprecated ACL::ShutdownFramework.
The protobuf message `ACL::ShutdownFramework` is marked as deprecated
since Mesos 0.27.0 in favor of the `ACL::TeardownFramework` message.
This patch prepares for the eventual removal of the Shutdown message
by limiting its usage to the definition of the ACLs, so that no
Mesos code actually uses the message.

Users can still define ACLs using the Shutdown message, in such
cases the following rules apply:

- If both Shutdown and Teardown are used, a warning message is
  printed informing of the deprecation and only the Teardown
  contents are used.
- If only Shutdown is used, a warning message is printed
  informing about deprecation, the contents of Shutdown are
  moved over to Teardown and Teardown contents are used.
- If only Teardown is used, no further action is taken.

Review: https://reviews.apache.org/r/44318/