Made changes to the executor library around managing connections.
This change makes the following modifications to the library:

\- Removes passing connection objects to \`defer\` callbacks as it can
sometimes lead to deadlocks around destruction in the same execution
context.
\- Introduced 3 additional states \`CONNECTING\`, \`SUBSCRIBING\` and
\`SUBSCRIBED\`. The \`CONNECTING\` state helps us in identifying if a
connection attempt is in progress while the latter two states allows
us to drop subscribe calls if one is already is in progress.
\- Creates a random \`connectionID\` to demarcate a new connection
instance and  allowing to discard a state connection attempt.
\- Changes around setting the recovery timeout timer only once.
This allows us to later discard the recoveryTimeout callback
if we connected with the agent at a later point of time.

Review: https://reviews.apache.org/r/44580/