Modified basic HTTP authenticator creator to accept realm.
To accommodate different authentication realms for the master and agent,
the default basic HTTP authenticator needs to accept its authentication
realm as a parameter. This patch adds this parameter and modifies the
HTTP authentication tests to validate it appropriately. A new test was
also added: `HttpAuthenticationTest.BasicWithoutRealm`.

Review: https://reviews.apache.org/r/44678/