Fix for docker containerizer not configuring CFS quotas correctly.
Note that this patch only handles the custom executor case. Command task
case is not handled yet due to some technical difficulty. Please see
details in comments of r33174.

It would be nice to refactor all this isolation code in a way that can
be shared between all containerizers, as this is basically just copied
from the CgroupsCpushareIsolator, but that's a much bigger undertaking.

Review: https://reviews.apache.org/r/33174/