Changed subprocess comment to reflect Error handling of parent Hooks.
Previously the Error behavior (in particular where the child process
is killed) was not consistent among Error cases.
We should establish the convention that the Hook itself returns an
error, while the subprocess code is resposible for killing the child
process in case of such Error.

Review: https://reviews.apache.org/r/44791/