Subprocess: [3/7] Introduced watchdog option.
Some newly created processes such as perf should be killed in case the
parent dies. Currently this is achieved by forking a new process from
the child process which serves as a 'watchdog' and kill the child if the
parent dies. This review introduces this as a general behavior into
subprocess (and hence removes the need for the custom setup function).

Review: https://reviews.apache.org/r/45232/