Added a script to install the Nvidia GDK on a host.
This script can be used to install the Nvidia GPU Deployment Kit (GDK)
on a mesos development machine. The purpose of the GDK is to provide
stubs for compiling code against Nvidia's latest drivers.  It includes
all the necessary header files (nvml.h) and library files
(libnvidia-ml.so) necessary to build mesos with Nvidia GPU support.
However, it does not actually contain any driver code, and it can't be
used to interact with any GPUs installed on a system. For that, you
will have to find the latest released drivers themselves.

We provide this script only for convenience, so that developers
working on non-GPU equipped systems can build code with Nvidia GPU
support. It is a simple wrapper around the official GDK installer
script provided by Nvidia, which can be downloaded here:

  https://developer.nvidia.com/gpu-deployment-kit

The script itself takes a few parameters. Run it as
'support/install-nvidia-gdk.sh --help' to see its usage semantics.

Review: https://reviews.apache.org/r/44360/