Added configure flags to build with Nvidia GPU support.
This is the initial commit to begin adding native support for GPUs in
Mesos. This initial version will only include support for Nvidia GPUs
that can be managed by the Nvidia Management Library (NVML).

The configure flags added in this commit can be used to enable Nvidia
GPU support, as well as specify the installation directories of the
NVML header and library files if not already installed in standard
include/library paths on the system.

In a subsequent commit, we will use these configure flags to
conditionally build support for Nvidia GPUs into Mesos.

Review: https://reviews.apache.org/r/44361/