Added stubs for the Nvidia GPU device isolator.
This is the first isolator to fall under the category of cgroup
devices. However, we do not yet have a generic cgroup device isolator
(nor will we in the very near future). As such, I have preemptively
created the nvidia gpu isolator in the directory hierarchy under the
path:

src/slave/containerizer/mesos/isolators/cgroups/devices/gpus/

in order to easily allow them to fall under the category of gpu
devices later on.

In this stub implementation, initialization of the agent will
fail if the nvidia gpu isolator is enabled via the agent --isolation
flag. That is --isolation="cgroups/devices/gpu/nvidia". In a
subsequent commit we will fill in the guts to actually enable the
proper isolation.

The flags documentation has been udpated accordingly.

Review: https://reviews.apache.org/r/44363/