Fix the local Docker puller for private registries.
If you use the unified containerizer and the local docker puller with a
Docker image from a private registry, the local puller fails to find any
layers in the image's repository manifest. This happens because the top
layer repository is qualified by the private registry name.

The fix is to also try the fully-qualified repository name if we have a
registry name.

Review: https://reviews.apache.org/r/45451/