Completed initial implementation of the Nvidia GPU isolator.
This isolator uses the cgroups devices subsystem to control access to Nvidia
GPUs. Since this is the very first device isolator, it currently contains
generic device isolation logic that needs to be pulled up into a generic device
isolator.

GPUs are allocated to containers in an arbitrary fashion.  For example, if a
container requires 2 GPUs, we will arbitrarily choose 2 from the GPUs that are
available.  This may not behave well if tasks within an executor use GPUs since
we cannot identify which task are using which GPUs (i.e. when a task
terminates, we may remove a GPU that is still being used by a different task!).

Note that this isolator is not responsible for ensuring that the necessary
Nvidia libraries are visible in the container. If filesystem isolation is not
enabled, this means that the container can simply use the libraries available
on the host. When filesystem isolation is enabled, it is the responsibility of
the operator / application developer to ensure that the necessary libraries are
visible to the container (note that they must be version compatible with the
kernel driver on the host).

TODO(klueska): To better support containers with a provisioned filesystem, we
will need to add a mechanism for operators to inject the libraries as a volume
into containers that require GPU access.

TODO(klueska): If multiple containerizers are enabled, they need to co-ordinate
their allocation of GPUs.

TODO(klueska): Move generic device isolation logic out into its own component.

This was joint work with Rob Todd from Nvidia <rtodd@nvidia.com>.

Review: https://reviews.apache.org/r/45439/